roulette

+++ client/src/components/currencyContext.jsx
@@ -4,7 +4,7 @@ import Cookies from 'js-cookie'
 const CurrencyContext = createContext();
 
 export const CurrencyProvider = ({ children }) => {
-    const [soulCoins, setSoulCoins] = useState(parseInt(Cookies.get("soulCoins")) || 0);
+    const [soulCoins, setSoulCoins] = useState(parseFloat(Cookies.get("soulCoins")) || 0);
     const convertToSoulCoins = (gold, silver, copper) => {
         const totalSoulCoins = (gold * 10) + (silver * 1) + (copper * 0.1);
         setSoulCoins(prev => prev + totalSoulCoins);
